# public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel Todo and MFA API",
        "version": "1.0.0",
        "description": "API documentation for Todo management and Multi-Factor Authentication"
    },
    "paths": {
        "/api/todos": {
            "get": {
                "summary": "Get all todos",
                "tags": ["Todos"],
                "responses": {
                    "200": {
                        "description": "List of all todos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "todos": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Todo"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new todo",
                "tags": ["Todos"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["title"],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "default": false
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Todo created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/todos/{id}": {
            "get": {
                "summary": "Get a specific todo",
                "tags": ["Todos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Todo ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            },
            "put": {
                "summary": "Update a todo",
                "tags": ["Todos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Todo ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "completed": {
                                        "type": "boolean"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Todo updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "todo": {
                                            "$ref": "#/components/schemas/Todo"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a todo",
                "tags": ["Todos"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Todo ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Todo deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Todo not found"
                    }
                }
            }
        },
        "/api/mfa/status": {
            "get": {
                "summary": "Get MFA status for authenticated user",
                "tags": ["MFA"],
                "responses": {
                    "200": {
                        "description": "MFA status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "enabled": {
                                            "type": "boolean"
                                        },
                                        "verified": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/setup": {
            "get": {
                "summary": "Setup MFA for authenticated user",
                "tags": ["MFA"],
                "responses": {
                    "200": {
                        "description": "MFA setup information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "secret_key": {
                                            "type": "string"
                                        },
                                        "qr_code_url": {
                                            "type": "string"
                                        },
                                        "recovery_codes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/enable": {
            "post": {
                "summary": "Enable MFA for authenticated user",
                "tags": ["MFA"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["code"],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "Verification code from authenticator app"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA enabled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "MFA setup not initialized"
                    },
                    "422": {
                        "description": "Invalid verification code"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/disable": {
            "post": {
                "summary": "Disable MFA for authenticated user",
                "tags": ["MFA"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["password"],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "description": "User's current password"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA disabled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "MFA is not enabled"
                    },
                    "422": {
                        "description": "Invalid password"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/verify": {
            "post": {
                "summary": "Verify MFA code",
                "tags": ["MFA"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["code", "type"],
                                "properties": {
                                    "code": {
                                        "type": "string",
                                        "description": "Verification code"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": ["totp", "recovery", "email"],
                                        "description": "Type of verification code"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "MFA verification successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "MFA is not enabled"
                    },
                    "422": {
                        "description": "Invalid verification code"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/email-code": {
            "post": {
                "summary": "Request email verification code",
                "tags": ["MFA"],
                "responses": {
                    "200": {
                        "description": "Email verification code sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "MFA is not enabled"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/mfa/recovery-codes": {
            "post": {
                "summary": "Regenerate recovery codes",
                "tags": ["MFA"],
                "responses": {
                    "200": {
                        "description": "Recovery codes regenerated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "recovery_codes": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "MFA not configured"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Todo": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "completed": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "MfaConfiguration": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "secret": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "recovery_codes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}